// apps/backend/src/modules/news-providers/models/normalized-article.interface.ts
import { NewsCategory } from '@prisma/client';

export interface NormalizedArticle {
  title: string;
  description: string | null;
  imageUrl: string | null;
  sourceName: string;
  sourceUrl: string;
  articleUrl: string;
  category: NewsCategory | null;
  publishedAt: Date | null;
  fullContent: string | null;
  provider: string;
}